Gameplay features
+ProjectileConfig scriptable object can be used to create different types of projectiles for the pitching machine to shoot out. This allows different projectiles to have different base speeds.

Pitcher has been updated to support an array of different pitchable objects. Pitcher config has been updated to supposed a speed modifier range for each pitch that's randomly selected at the time the object is thrown.

The ball projectile deals damage when it hits the player body but not the fist.

The fireball projectile deals damage when punched or when it hits the body. The purpose of the fireball is to be jumped over and avoided.

RobotAnimationEventHandler was updated to remove all instance of SendMessage. This was used early to get animation events moving across different components but it's slow and message and the messages were replaced by directly referenced function calls in the robotcontroller.